twitter-crawlers/requirements.txt
elasticsearch==6.3.0
tweepy==3.6.0
python_dateutil==2.7.3